refactored the genMoveRange method so it uses pathnodes to keep track of where its been (there by making it much more efficent and faster). generating move ranges of units becomes much more managable (i.e. taking at most 1 sec. instead of 5 and more for movements greater than 10 x.x). units seem to be moving to their maximum potential more often, and are on a whole being more effective.
This refactoring session has been very productive and helped the overall
quality of the project, as well as fixing some long outstanding bugs.
next agenda is to implement capital-pathfinding (so as to not be blocked
by units).

I can smell it. It's getting closer and closer to being completed. since
Sept. 2012 I've been working on this, and it's only getting better. It
just goes to show, never give up.